+++ src/itfPanel.c
@@ -3493,7 +3493,7 @@ u64 FUN_003acda0(u64 param_1,int param_2)
   else {
     uVar3 = (u32)(u8)(bVar2 - 1);
   }
-  packed = uVar3 << 8 | low;
+  packed = (u32)(u8)uVar3 << 8 | (u32)(u8)low;
   lVar4 = FUN_0016f190(0x184);
   if ((lVar4 != 0) && ((DAT_007cd500 & 0x100) == 0)) {
     FUN_0010a2e0(2,2,packed);
